Added the colours and a lot of feedback received
I added all the colours in the element and created a little function to be implemented on click. It uses state to change the output.
I added one class which takes the current colour and adds all the rest of the colours (border, shadows, outline, text, etc). I had to add separate classes for the void and primary as they do not use the actual colour grid.
I made the component reusable
I fixed all the feedback from Euan and it is all working now.
Just left to do some final bits and ensure that the elements can be reused. Will see if I can make the void component part of the main one as I had to create a new one for the void because it has some different behaviour