Added the BanIcon component to the main one
I added the component to the main one and took advantage of its classes and values. Just added the different ones in.
Also added a condition in the jsx which will check what the colour is and output the correct version.
I kept the old files for reference but now ColourThemeSwatch should work on its own for all colours