Made a reusable icon for tooltip component
I created a tooltip component and created prop which is dynamic and can change the icon that needs to be clicked/hovered.
I added the prop with capital letters as I will use it as a component too. After that I added the icon name in curly braces and then took this prop in the main component and added the name of the new component to be the prop name - Icon.