Tag component
I added 3 new components:
 - Tag
 - AddTag
 - AddRemoveBtnComponent
AddRemoveBtnComponent is just a button with all its properties and effects that it reusable and can be added anywhere as such.
Tag is the tag component I created. It had children prop for its text and integrated AddRemoveBtnComponent which can be adjusted with props from the parent component. It has a lot of different props in order to make it as customisable as possible.
AddTag component is another component which takes AddRemoveBtnComponent as a child and can change its properties as well as some other options.