Added a modal component
I created a modal component template. It has the basics structure of a modal with a function to close it on click on the X or cancel buttons.
I added the infoText in another component and it can be tailored with props - children for the text and color for the color needed. I included InfoText component in the Modal component and made its props dynamic so they can be passed to the parent.
Modal still needs some improvement in the buttons as I have not added the icon yet but as I used the Button component I believe we should either add a prop for icon there or I am missing it somewhere, so I will wait to ask Rory.