I made the children prop of InfoText more dynamic
I removed some of the JSX in the InfoText component and left just the children prop in a div. It would make it more dynamic as I spoke to Euan and there are a few different boxes with inline heading or icon. So it would be better to add them as props in the instance of the parent component.