Added a function
I added a little function to count the characters of the children elements and then added some conditional to change the width of the element based on the characters amount for left/right.
I also made the height fixed so that the tooltip only widens based on the characters and height stays fixed to keep the arrow on the same place.
Also had to adjust the left side tooltip as it got displaced because I removed the width of the parent element.