Modal fixes
I removed the fixed width and added % plus max-width for the container
I removed the height from the container as well as the children container. So the height now is going to be determined by the content
Changed the span to button (X icon) in order to make it close the modal on enter key when focused
Added fixed min-height of 60px to the title container
Fixed the height of the modal. I made it have max-height after which it will take overflow: auto. If not enough content is provided the modal will have the height of the content
I made the button Continue dynamic - added props btnText, BtnIcon, handleClick and added a conditional to show the icon only if BtnIcon is true because I get error if nothing is passed. Now it keeps the text only if no icon is passed
Added border-bottom to the children element and fixed the padding and margin around it