Added another prop to IconTooltip
I added another prop to IconTooltip to make the width of the tooltip flexible as I could not find another way to make it follow the content width. It is set by default to ensure that it works for normal and long text but was not the best for very short text. Now user can adjust the width of the element based on the amount of text added