Created a new element
I created a new element where I added the HexColorInput one. This is an element with an input field to type a hex value
I also added another function to the ColourPicker component which calculated a border colour based on the variable color value. I made it a bit darker than the actual color